chore: move const file to common directory, modify content import path
